Make bucket graph
